Changing configuration structure to enable multiple accounts.
This change enables using multiple wire plugins at the same time.
Also, we now distinguish between the wire plugin (i.e. EBICS or
taler_bank) and the wire method (i.e. SEPA or x-taler-bank) that
the wire plugin is implementing.  The "taler-bank" wire method
was renamed from "test" to "x-taler-bank".

This also changes the format of the /wire response of the exchange,
as we now need to return multiple accounts.  Note that wire fees
are specified per wire method, not per wire account.

taler-exchange-keyup now automatically signs all of the /wire
responses in the location specified by the configuration.

Account identification in wire plugins was changed to use
payto://-URLs instead of method-specific JSON fields.  Signing
and validation of /wire responses was moved from each wire
plugin to a generic validation method in libtalerutil (crypto)
or libtalerjson (for JSON-formatted inputs).
Convenience methods were added to generate JSON for wire accounts
(salting, signing).

Various section and option names were adjusted to streamline the
configuration and make it more consistent overall.  Documentation
was updated as well.

+++ src/bank-lib/testing_api_cmd_reject.c
@@ -127,7 +127,7 @@ reject_run (void *cls,
     (GNUNET_OK == TALER_TESTING_GET_TRAIT_ROW_ID
       (deposit_cmd, &row_id));
   TALER_LOG_INFO ("Account %llu rejects deposit\n",
-                  *credit_account);
+                  (unsigned long long) *credit_account);
   rs->rh = TALER_BANK_reject (is->ctx,
                               rs->bank_url,
                               &AUTHS[*credit_account -1],
@@ -180,7 +180,7 @@ TALER_TESTING_cmd_bank_reject (const char *label,
   rs = GNUNET_new (struct RejectState);
   rs->bank_url = bank_url;
   rs->deposit_reference = deposit_reference;
-  
+
   cmd.cls = rs;
   cmd.run = &reject_run;
   cmd.cleanup = &reject_cleanup;
